+++ src/sll/vm.c
@@ -76,7 +76,7 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_return_code_t sll_execute_assembly(const s
 	sll_current_instruction_count=0;
 	sll_current_vm_config=cfg;
 	const sll_assembly_instruction_t* ai=a_dt->h;
-	sll_page_size_t ptr_sz=SLL_ROUND_LARGE_PAGE(cfg->s_sz+a_dt->vc*sizeof(sll_object_t*)+a_dt->st.l*sizeof(sll_object_t)+SLL_CALL_STACK_SIZE*sizeof(sll_call_stack_frame_t));
+	sll_size_t ptr_sz=SLL_ROUND_LARGE_PAGE(cfg->s_sz+a_dt->vc*sizeof(sll_object_t*)+a_dt->st.l*sizeof(sll_object_t)+SLL_CALL_STACK_SIZE*sizeof(sll_call_stack_frame_t));
 	uint64_t ptr=(uint64_t)sll_platform_allocate_page(ptr_sz,1);
 	sll_object_t** v=(sll_object_t**)ptr;
 	sll_static_int[0]->rc+=a_dt->vc;